Move business logic from controllers into service and action classes

+++ app/Http/Controllers/PostController.php
@@ -2,12 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Actions\GenerateThumbnail;
 use App\Models\Post;
+use App\Services\PostService;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-use Intervention\Image\ImageManagerStatic as Image;
 
 class PostController extends Controller
 {  
@@ -31,7 +31,7 @@ class PostController extends Controller
      * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function store(Request $request): RedirectResponse
+    public function store(Request $request, PostService $postService, GenerateThumbnail $generateThumbnail): RedirectResponse
     {
 		$validate = $request->validate([
 			'post_title' => 'required|max:255',
@@ -39,29 +39,12 @@ class PostController extends Controller
 			'post_image' => 'required|image|mimes:png,jpg,bmp,gif'
 		]);
 
-        if (!$request->file('post_image'))
-			return redirect()
-				->route('posts.create')
-				->with('message', 'no_image');
-		
-		// Upload image
-		$image = $request->file('post_image');
-		$fileName = date('d_m_Y_H_i') . $image->getClientOriginalName();
-		$thumbnail = 'thumbnail_' . $fileName;
-		$image->storeAs('public/uploads', $fileName);
-		$image->storeAs('public/uploads/thumbnails', $thumbnail);
-		
-		// Generate thumbnail
-		$thumbnailPath = public_path('storage/uploads/thumbnails/' . $thumbnail);
-		$this->generateThumbnail($thumbnailPath, 368, 240);
-
-		$post = new Post();
-		$post->title = $request->input('post_title');
-		$post->content = $request->input('post_content');
-		$post->image_path = $fileName;
-		$post->thumbnail_path = $thumbnail;
-		$post->user_id = $request->user()->id;
-		$post->save();
+        $post = $postService->createPost(
+			$request->input('post_title'),
+			$request->input('post_content'),
+			$request->file('post_image'),
+			$generateThumbnail
+		);
 
 		return redirect()
 			->route('posts.show', ['post' => $post])
@@ -87,6 +70,8 @@ class PostController extends Controller
      */
     public function edit(Post $post): View
     {
+		$this->authorize('update', $post);
+
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -97,38 +82,22 @@ class PostController extends Controller
      * @param  \App\Models\Post  $post
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function update(Request $request, Post $post): RedirectResponse
+    public function update(Request $request, Post $post, PostService $postService, GenerateThumbnail $generateThumbnail): RedirectResponse
     {
+		$this->authorize('update', $post);
+
         $validate = $request->validate([
 			'post_title' => 'min:6|max:255',
 			'post_image' => 'image|mimes:png,jpg,bmp,gif'
 		]);
-		
-		// Only update the fields in the database if they have been changed
-		if ($post->title !== $request->input('post_title'))
-			$post->title = $request->input('post_title');
-		
-		if ($post->content !== $request->input('post_content'))
-			$post->content = $request->input('post_content');
 
-		if ($request->file('post_image'))
-		{
-			$image = $request->file('post_image');
-			$fileName = date('d_m_Y_H_i') . $image->getClientOriginalName();
-			$thumbnail = 'thumbnail_' . $fileName;
-			$image->storeAs('public/uploads', $fileName);
-			$image->storeAs('public/uploads/thumbnails', $thumbnail);
-			
-			// Generate thumbnail
-			$thumbnailPath = public_path('storage/uploads/thumbnails/' . $thumbnail);
-			$this->generateThumbnail($thumbnailPath, 368, 240);
-
-			$post->image_path = $fileName;
-			$post->thumbnail_path = $thumbnail;
-		}
-
-		$post->updated_by = Auth::user()->id;
-		$post->save();
+		$postService->updatePost(
+			$post->id, 
+			$request->input('post_title'), 
+			$request->input('post_content'), 
+			$request->file('post_image'),
+			$generateThumbnail
+		);
 
 		return redirect()
 			->route('posts.show', ['post' => $post])
@@ -143,24 +112,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+		$this->authorize('delete', $post);
         $post->delete();
 		
 		return redirect()
 			->route('posts.index')
 			->with('action', 'post_deleted');
     }
-
-	/**
-	 * Generate a thumbnail of specified size
-	 *
-	 * @param  string $path path of thumbnail
-	 * @param  int $width
-	 * @param  int $height
-	 * @return void
-	 */	
-	public function generateThumbnail(string $path, int $width, int $height): void
-	{
-    	$img = Image::make($path)->resize($width, $height)
-			->save($path);
-	}
 }
